Informed `figwheel` about CSS file (???)
...But I do not really know if this change has caused any effect. I
tried changing a number of settings, but saw no visible change. However,
I do not know if the lack of change resulted from an error in my setup
or from not understanding all the CSS details.

Well, I was able to verify that changing the `app` `color` property
resulted in changing the value in the element inspector (although it was
struck through - which I think indicates that it is overridden somehow).
In addition, I did observe that whenever I saved changes to the site.css
file, the debug console displayed the message '("public/css/site.css")'.

I think these changes indicate everything is configured correctly;
however, I am not 100% confident.