Correct errors entering (some) URLs
Previous commits raised errors when entering or refreshing some URLs. (I
believe it choked on refreshing URLs like "" or "/").

This commit corrects many of those errors; however, entering "/about",
which to the user would seem to be valid, causes FigWheel (and python
`SimpleHTTPServer`) to display an error message. I'm not certain how to
address this. I hope to address it at some point.

I have *so much* to learn...