Repair error it active page sytling
The video used a feature of `react-router` to indicate a page was
active: it displayed the link for that page in the header with
class="active" and added styling for `a.active`. Previous commits *did
not* exhibit this same behavior.

This commit makes the behavior identical - although it is a bit odd (but
the "oddness" is in the original).